+++ src/app/carrier/carrier-view/carrier-view.component.ts
@@ -77,14 +77,29 @@ export class CarrierViewComponent implements OnInit {
     this.router.navigate(['home/dashboard']);
   }
 
+  selectedRows: any = []
+  onChangeSelection(event) {
+    console.log(event)
+    this.selectedRows = event;
+  }
+
   OnAddClick() {
-    localStorage.setItem("DD_ROW", "");
-    localStorage.setItem("Action", "");
-    this.carrierMainComponent.carrierComponent = 2;
+    if (this.selectedRows.length > 0) {
+      this.event = event;
+      this.dialogFor = "DataLost";
+      this.yesButtonText = this.translate.instant("yes");
+      this.noButtonText = this.translate.instant("no");
+      this.showConfirmDialog = true;
+      this.dialogMsg = this.translate.instant("SelectionLostMsg");
+    } else {
+      localStorage.setItem("DD_ROW", "");
+      localStorage.setItem("Action", "");
+      this.carrierMainComponent.carrierComponent = 2;
+    }
   }
 
   OnDeleteSelected(event) {
-    if(event.length <= 0){
+    if (event.length <= 0) {
       this.toastr.error('', this.translate.instant("CAR_deleteitem_Msg"));
       return;
     }
@@ -149,6 +164,11 @@ export class CarrierViewComponent implements OnInit {
     this.showConfirmDialog = false;
     if ($event.Status == "yes") {
       switch ($event.From) {
+        case ("DataLost"):
+          localStorage.setItem("DD_ROW", "");
+          localStorage.setItem("Action", "");
+          this.carrierMainComponent.carrierComponent = 2;
+          break
         case ("Delete"):
           var ddDeleteArry: any[] = [];
           ddDeleteArry.push({
